requirements.txt
requests
openai
pydantic
python-dotenv
fastapi
uvicorn
jinja2
python-multipart
voyageai
numpy
